Change everything back to storing insert IDs, not pointers
because pointers were a *stupid* idea. That is, you see, because sometimes realloc() moves the memory for the insert set. So the stored insert pointer is still pointing to the same place, but the insert it was supposed to be pointing to is *gone*. m( It is exactly the kind of bug that you don't want, because it *only* occurs when we need to realloc some more insert memory, *and* realloc decides to move our memory around, so it's like "oh, the program segfaulted. Let's try reproducing the bug under valgrind. *nothing happens* ...".