Rewrite the send_queue, insert resending
Remove network_data's send_queue_free_slots (which could have caused a segfault anyway), and use send_queue just as a normal array. On enqueuing, insert IDs are now appended onto the send_queue, and every 500ms (for now) the first element of the send_queue is popped of, resent, and stuck on the back. On ACK, the insert ID is just deleted out of the queue.