Add predictive character insertion and deletion to main.c
Instead of merely waiting for the update from the Rust thread, the C thread now inserts the character as though it was a normal editor, and synchronizes later. Due to the way the backend insert works, there should be no discrepancy between these two versions, and the synchronization should go by unnoticed.